Refactored some cruft out of Sudoku into a GridGeneratingFactory.
