+++ HW3-SudokuSolver/build.gradle
@@ -8,7 +8,8 @@ repositories {
 
 dependencies {
     compile group: 'commons-collections', name: 'commons-collections', version: '3.+'
+    compile 'org.apache.commons:commons-lang3:3.+'
     testCompile group: 'junit', name: 'junit', version: '4.+'
 }
 
-mainClassName = "assign3.Sudoku"
+mainClassName = "assign3.Main"
