+++ HW3-SudokuSolver/src/main/java/assign3/Sudoku.java
@@ -25,60 +25,6 @@ public class Sudoku {
 	private List<HashSet<Integer>> valInRows, valInCols, valInParts;
 	
 	private long timeTakenForSolution;
-	
-	/* Parts of the grid each of size 3x3. Counting from the
-	 * top left to top right then the next row below.
-	 * 0 1 2
-	 * 3 4 5
-	 * 6 7 8
-	 */
-	private static final int PART1 = 0;
-	private static final int PART2 = 1;
-	private static final int PART3 = 2;
-	private static final int PART4 = 3;
-	private static final int PART5 = 4;
-	private static final int PART6 = 5;
-	private static final int PART7 = 6;
-	private static final int PART8 = 7;
-	private static final int PART9 = 8;
-		
-	// Provided easy 1 6 grid
-	public static final int[][] easyGrid = Sudoku.stringsToGrid(
-	"1 6 4 0 0 0 0 0 2",
-	"2 0 0 4 0 3 9 1 0",
-	"0 0 5 0 8 0 4 0 7",
-	"0 9 0 0 0 6 5 0 0",
-	"5 0 0 1 0 2 0 0 8",
-	"0 0 8 9 0 0 0 3 0",
-	"8 0 9 0 4 0 2 0 0",
-	"0 7 3 5 0 9 0 0 1",
-	"4 0 0 0 0 0 6 7 9");
-	
-	
-	// Provided medium 5 3 grid
-	public static final int[][] mediumGrid = Sudoku.stringsToGrid(
-	 "530070000",
-	 "600195000",
-	 "098000060",
-	 "800060003",
-	 "400803001",
-	 "700020006",
-	 "060000280",
-	 "000419005",
-	 "000080079");
-	
-	// Provided hard 3 7 grid
-	// 1 solution this way, 6 solutions if the 7 is changed to 0
-	public static final int[][] hardGrid = Sudoku.stringsToGrid(
-	"3 7 0 0 0 0 0 8 0",
-	"0 0 1 0 9 3 0 0 0",
-	"0 4 0 7 8 0 0 0 3",
-	"0 9 3 8 0 0 0 1 2",
-	"0 0 0 0 4 0 0 0 0",
-	"5 2 0 0 0 6 7 9 0",
-	"6 0 0 0 2 1 0 4 0",
-	"0 0 0 5 3 0 9 0 0",
-	"0 3 0 0 0 0 0 5 1");
 
 	/* Updates the Spot on the solution grid with the row and col of
      * this Spot with the current value that this spot holds.
@@ -87,29 +33,8 @@ public class Sudoku {
 		solutionGrid[spot.getRow()][spot.getCol()] = spot;
 	}
 	
-	public static final int SIZE = 9;  // size of the whole 9x9 puzzle
-	public static final int PART = 3;  // size of each 3x3 part
+	public static final int SIZE = 9;  // size of the sudoku puzzle, always a perfect square
 	public static final int MAX_SOLUTIONS = 100;
-	
-	// Provided various static utility methods to
-	// convert data formats to int[][] grid.
-	
-	/**
-	 * Returns a 2-d grid parsed from strings, one string per row.
-	 * The "..." is a Java 5 feature that essentially
-	 * makes "rows" a String[] array.
-	 * (provided utility)
-	 * @param rows array of row strings
-	 * @return grid
-	 */
-	public static int[][] stringsToGrid(String... rows) {
-		int[][] result = new int[rows.length][];
-		for (int row = 0; row<rows.length; row++) {
-			result[row] = stringToInts(rows[row]);
-		}
-		return result;
-	}
-
 
 	/**
 	 * Given a single string containing 81 numbers, returns a 9x9 grid.
@@ -121,7 +46,7 @@ public class Sudoku {
 	public static int[][] textToGrid(String text) {
 		int[] nums = stringToInts(text);
 		if (nums.length != SIZE*SIZE) {
-			throw new RuntimeException("Needed 81 numbers, but got:" + nums.length);
+			throw new RuntimeException(String.format("Needed %d numbers, but got : %d", SIZE*SIZE, nums.length));
 		}
 		
 		int[][] result = new int[SIZE][SIZE];
@@ -134,7 +59,6 @@ public class Sudoku {
 		}
 		return result;
 	}
-	
 	
 	/**
 	 * Given a string containing digits, like "1 23 4",
@@ -400,17 +324,4 @@ public class Sudoku {
 		}
 		return result;
 	}
-	
-	/* Just for simple testing */
-	public static void main(String[] args) {
-		Sudoku sudoku;
-		sudoku = new Sudoku(hardGrid);
-		
-		System.out.println(sudoku); // print the raw problem
-		int count = sudoku.solve();
-		System.out.println("solutions:" + count);
-		System.out.println("elapsed:" + sudoku.getElapsed() + "ms");
-		System.out.println(sudoku.getSolutionText());
-	}
-
 }
